Multicolour sets
Mostly all four colours of eyelashes, with a few multiple colours of
accessories. Japanese can fit this all on one line, with one character
per colour. Another solution might be to put the four colours on the
next line in brackets, but that could run into problems if they later
make a set with two other items on top of the eyelashes - they already
have a set with one other item.