Hien-tsubaki chip ability name
The tsubaki/camellia flower is an omen of bad luck to samurai, as it
drops off all at once instead of petal by petal, and its red colour
reminded them of decapitated heads falling. I tried to come up with
something poetic that conveyed the same kind of thing to translate
"Tsubaki" here.